add (core: platform/allocators) -> added alloc virtual mem - added a start for platform allocator - fixed windows exit() - added shell dcript to build windows

+++ include/seul/platform.h
@@ -55,7 +55,7 @@ i32 seul_platform_windows_syscall(usize syscall_number, ...);
 
 
 #if __target_openbsd
-i32 seul_platform_openbsd_syscall(usize syscall_number, ...);
+usize seul_platform_openbsd_syscall(usize syscall_number, ...);
 #endif
 
 // NOTE: platform generic defs
